fix: 상세페이지 북마크 적용완료
