+++ client/src/pages/Category/Category.js
@@ -111,33 +111,7 @@ export default function Category() {
       totalPages: 5,
     });
   }, [fitlerCondtion, currentPage]);
-  const handleBookmarkClick = (cocktailId, item) => {
-    const id = cocktailId;
-    dispatch(updateBookmark({ id, item }));
 
-    const newDate = data.map((el, idx) => {
-      if (idx + 1 === item.cocktailId) {
-        return { ...el, isBookmarked: !el.isBookmarked };
-      }
-      return el;
-    });
-    setData(newDate);
-    // /cocktails/{cocktail-id}/bookmark
-    // const handleBookmark = () => {
-    //   fetch(`/cocktails/${item.cocktailId}/bookmark`, {
-    //     method: 'POST',
-    //     // 필요한 경우 헤더 등을 설정하세요.
-    //   })
-    //     .then((response) => {
-    //       if (!response.ok) {
-    //         throw new Error('Bookmarking failed.'); // 요청이 실패한 경우 에러 처리
-    //       }
-    //       // 요청이 성공한 경우 추가적인 작업을 수행할 수 있습니다.
-    //     })
-    //     .catch((error) => {
-    //       console.error(error); // 에러 처리
-    //     });
-  };
   return (
     <div className="overflow-hidden">
       <div className="relative bg-gradient-to-r from-gradi-to to-gradi-from w-screen h-100% pt-[10rem] flex justify-center  ">
@@ -184,8 +158,9 @@ export default function Category() {
                 <Card
                   item={item}
                   className="pr-4"
-                  key={index}
-                  onClick={() => handleBookmarkClick(item.cocktailId, item)}
+                  key={index + 1}
+                  data={data}
+                  setData={setData}
                 />
               ))}
             </div>
