+++ client/src/pages/RecipeDetail/Recommend.js
@@ -7,17 +7,17 @@ import { useState } from 'react';
 
 export default function Recommend({ cocktailDetail }) {
   const [recommends, setRecommends] = useState(cocktailDetail);
+  const handleBookmarkClick = () => {};
   return (
     <RecommendContainer>
       <RecommentP>이런 칵테일은 어떠세요?</RecommentP>
       <CardContainer>
         {recommends.map((ele) => {
           return (
-            <Link to={`/detail/${ele.cocktailId}`} key={ele.cocktailId}>
-              <div className="mb-12 mr-4">
-                <Card item={ele} setData={setRecommends} data={recommends} />
-              </div>
-            </Link>
+            //북마크 클릭시 link 가 이동해서 card 안에 navigate(`/detail/${item.cocktailId}`)으로 수정했습니다.
+            <div className="mb-12 mr-4" key={ele.cocktailId}>
+              <Card item={ele} setData={setRecommends} data={recommends} />
+            </div>
           );
         })}
       </CardContainer>
